requirements.txt
streamlit>=1.24.0
ollama>=0.1.0
numpy>=1.21.0
pandas>=1.3.0
PyPDF2>=3.0.0
python-docx>=0.8.11
scikit-learn>=1.0.0
Pillow>=9.0.0